add confirmation to copy and delete from history

+++ src/main/java/hu/adamsan/utilities/filecomparer/ui/DifferenceController.java
@@ -8,7 +8,10 @@ import org.apache.commons.io.FileUtils;
 import hu.adamsan.utilities.filecomparer.Pair;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TitledPane;
@@ -78,7 +81,6 @@ public class DifferenceController {
     public void copySelected(ActionEvent event) {
         MyCellData selected = differenceTable.getSelectionModel().getSelectedItem();
         if (selected != null) {
-            System.out.println("TODO: Copy selected");
             String selectedFileString = selected.getName().get();
             copy(selectedFileString);
             mainUI.compare(currentPair);
@@ -89,15 +91,30 @@ public class DifferenceController {
     private void copy(String selectedFileString) {
         File source = new File(currentPair.first, selectedFileString);
         File dest = new File(currentPair.second, selectedFileString);
-        try {
-            if (source.isDirectory()) {
-                FileUtils.copyDirectory(source, dest);
-            } else {
-                FileUtils.copyFile(source, dest);
+
+        if (confirmCopy(source, dest)) {
+
+            try {
+                if (source.isDirectory()) {
+                    FileUtils.copyDirectory(source, dest);
+                } else {
+                    FileUtils.copyFile(source, dest);
+                }
+            } catch (IOException e) {
+                e.printStackTrace();
             }
-        } catch (IOException e) {
-            e.printStackTrace();
         }
+    }
+
+    private boolean confirmCopy(File source, File dest) {
+        Alert alert = new Alert(AlertType.CONFIRMATION);
+        alert.setTitle("Confirm copy");
+        alert.setContentText("Are you sure, you want to copy \n" +
+                source + "\n to \n" +
+                dest + " ?");
+        alert.showAndWait();
+        ButtonType result = alert.getResult();
+        return result == ButtonType.OK;
     }
 
     public void setCurrentPair(Pair<File> currentPair) {
